Project/requirements.txt
absl-py==2.1.0
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
astunparse==1.6.3
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cryptography==42.0.2
cssselect2==0.7.0
cycler==0.12.1
Django==4.2
django-cleanup==8.0.0
flatbuffers==24.3.25
fonttools==4.47.2
gast==0.5.4
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.60.1
h5py==3.10.0
html5lib==1.1
idna==3.6
joblib==1.3.2
keras==2.15.0
kiwisolver==1.4.5
libclang==16.0.6
libretranslatepy==2.1.1
lxml==5.1.0
Markdown==3.5.2
MarkupSafe==2.1.5
matplotlib==3.8.2
ml-dtypes==0.2.0
mysqlclient==2.2.1
numpy==1.26.3
oauthlib==3.2.2
opt-einsum==3.3.0
oscrypto==1.3.0
packaging==23.2
pandas==2.1.4
pillow==10.2.0
protobuf==4.23.4
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pyHanko==0.21.0
pyhanko-certvalidator==0.26.3
pyparsing==3.1.1
pypdf==4.0.1
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
qrcode==7.4.2
reportlab==4.0.9
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.12.0
seaborn==0.13.2
six==1.16.0
sqlparse==0.4.4
svglib==1.5.1
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.2.0
tinycss2==1.2.1
translate==3.6.1
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
uritools==4.0.2
urllib3==2.2.0
webencodings==0.5.1
Werkzeug==3.0.1
wrapt==1.14.1
xhtml2pdf==0.2.15
